requirements.txt
requests
django
django-filter
djangorestframework
urllib3==1.26.15
mongoengine
gunicorn
psycopg2-binary
django-pymemcache
python-memcached